refactor: pull apply/hydrate logic out of writer
