+++ packages/translator/src/core/condition/else.ts
@@ -1,6 +1,7 @@
 import type { types as t } from "@marko/compiler";
 import { Tag, assertNoParams, assertNoVar } from "@marko/babel-utils";
 import * as writer from "../../util/writer";
+import { setQueueFactory } from "../../util/apply-hydrate";
 import { exitCondition, queueBranchFactory } from "./if";
 
 export default {
@@ -32,7 +33,7 @@ export default {
       }
 
       writer.start(tag);
-      writer.setQueueFactory(tag, queueBranchFactory);
+      setQueueFactory(tag, queueBranchFactory);
     },
     exit(tag) {
       exitCondition(tag);
